Add implementation for shell env, bug fixes

+++ src/main/java/game/actions/Action.java
@@ -3,5 +3,5 @@ package game.actions;
 public interface Action {
 	public ActionType getActionType();
 	public long getMoveValue();
-	public long getRelativeMoveValue();
+	public float getRelativeMoveValue();
 }
